MELQ-171: Introduce PostContent
